Extract UITestAPIStore actor out of APIClient (#416)
Move the UI-test infrastructure (UITestConfig, UITestStore + makeDefault/
makeSnapshot, and the uiTest* request handlers + backing stored props) out of
APIClient into a dedicated UITestAPIStore actor. APIClient now keeps a single
seam: an optional uiTestAPIStore built from the same SP_UI_TEST_* env vars, and
delegates to it whenever it is non-nil.

clearLocalSessionArtifacts() stays owned by APIClient. Instead of injecting a
callback, logout()/deleteAccount() on the store return a flag telling the client
to clear its session artifacts, and the store exposes resetStore so the client
can clear them during init on a reset launch.
